+++ resources/js/lang/en/app.js
@@ -46,10 +46,7 @@ export default {
 
   footer: {
     legalNotice: 'Legal notice',
-    privacyPolicy: 'Privacy policy',
-    legalNoticeUrl: 'https://www.thm.de/ges/en/about',
-    privacyPolicyUrl: 'https://www.thm.de/ges/en/datenschutz',
-    separator: '|'
+    privacyPolicy: 'Privacy policy'
   },
 
   roles: {